--- src/main/java/com/daangn/dangunmarket/domain/member/member/domain/MemberProvider.java
@@ -1,7 +1,0 @@
-package com.daangn.dangunmarket.domain.member.member.domain;
-
-public enum MemberProvider {
-    KAKAO,
-    GOOGLE
-    ;
-}
