+++ src/main/java/com/daangn/dangunmarket/domain/member/model/MemberProvider.java
@@ -1,0 +1,7 @@
+package com.daangn.dangunmarket.domain.member.model;
+
+public enum MemberProvider {
+    KAKAO,
+    GOOGLE
+    ;
+}
