+++ src/main/java/com/daangn/dangunmarket/domain/member/model/NickName.java
@@ -1,4 +1,4 @@
-package com.daangn.dangunmarket.domain.member.member.domain;
+package com.daangn.dangunmarket.domain.member.model;
 
 import jakarta.persistence.Embeddable;
