+++ src/main/java/com/daangn/dangunmarket/domain/member/model/Review.java
@@ -1,6 +1,5 @@
-package com.daangn.dangunmarket.domain.member.review.domain;
+package com.daangn.dangunmarket.domain.member.model;
 
-import com.daangn.dangunmarket.domain.member.member.domain.Member;
 import jakarta.persistence.*;
 import lombok.*;
